Removing the isDefined check for scope.resultImage
Removing the isDefined check for scope.resultImage. I should not be required to define the property I bind to the resultImage object to. This is a very unnecessary requirement to place upon your users and it can cause issues when using complex javascript objects which do not have their entire structure predefined in code.